Add files via upload
For educational purposes only, this project will guide you through the steps on how to exploit vulnerability on the LinESC machine.

The aim is to provide an understanding of the vulnerabilities within the LinESC operating system and to emphasize the importance of securing your system against unauthorized access. By following this project, you'll gain valuable insights into how these security mechanisms work, but remember, this is strictly for learning and should not be used for any malicious intent.